seperated the interface from the implementation, now only to add safety guarantees as we still allow double event with same id
